+++ Framework/Nexus/inc/MantidNexus/NexusFile.h
@@ -34,8 +34,6 @@ class H5Object;
 namespace Mantid {
 namespace Nexus {
 
-static Entry const EOD_ENTRY(NULL_STR, NULL_STR);
-
 /**
  * \class FileID
  * \brief A wrapper class for managing HDF5 file handles (hid_t).
@@ -96,10 +94,6 @@ private:
    * - firstEntryNameType
    */
   NexusDescriptor m_descriptor;
-  /** NOTE: this is temporary until `napi` is fully deleted*/
-  NexusAddress m_group_address;
-  /** NOTE: this is temporary until `napi` is fully deleted*/
-  hid_t m_fid;
 
   //------------------------------------------------------------------------------------------------------------------
   // CONSTRUCTORS / ASSIGNMENT / DECONSTRUCTOR
@@ -170,12 +164,6 @@ private:
   //------------------------------------------------------------------------------------------------------------------
 public:
   // ADDRESS GET / OPEN
-
-  /**
-   * DO NOT USE THIS FUNCTION FOR ANY REASON
-   * It is needed as a temporary solution while deleting napi.
-   */
-  NexusFile5 getFileStruct();
 
   /**
    * Open the NeXus object with the address specified.
